requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
flask-talisman>=0.8.1
```
 

 